server/requirements.txt
Authlib == 0.15.*
bcrypt == 3.*
Flask == 1.*
Flask-Cors == 3.*
loginpass == 0.5
pyjwt == 2.*
SQLAlchemy == 1.*
# SQLAlchemy == 1.4.0b1
